Reorganize File structure
