+++ Weekand/Weekand/Presentation/Common/Data/ColorSet.swift
@@ -1,0 +1,19 @@
+//
+//  Color.swift
+//  Weekand
+//
+//  Created by Daegeon Choi on 2022/05/20.
+//
+
+import UIKit
+
+extension UIColor {
+    // MARK: Custom System Colors
+    static let mainColor = UIColor(hex: "#5086FF")
+    static let backgroundColor = UIColor.white
+    
+    static let accentColor = UIColor(hex: "#17191C")
+    static let lightGray = UIColor(hex: "#ECEFF2")
+    static let darkGray = UIColor(hex: "#AAAEB6")
+    
+}
