+++ Weekand/Weekand/Presentation/SignScene/SignIn/View/SignInViewController.swift
@@ -1,0 +1,29 @@
+//
+//  SignInViewController.swift
+//  Weekand
+//
+//  Created by Daegeon Choi on 2022/05/20.
+//
+
+import UIKit
+
+class SignInViewController: ViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
