+++ Weekand/Weekand/Presentation/SignScene/SignIn/ViewModel/SignInViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  SignInViewModel.swift
+//  Weekand
+//
+//  Created by Daegeon Choi on 2022/05/20.
+//
+
+import Foundation
