+++ Weekand/Weekand/Presentation/SignScene/SignUp/View/SignUpViewController.swift
@@ -1,0 +1,29 @@
+//
+//  SignUpViewController.swift
+//  Weekand
+//
+//  Created by Daegeon Choi on 2022/05/20.
+//
+
+import UIKit
+
+class SignUpViewController: ViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
