+++ Weekand/Weekand/Presentation/Utils/UIColor+hex.swift
@@ -8,18 +8,7 @@
 import UIKit
 
 extension UIColor {
-    
-    // MARK: Custom System Colors
-    static let mainColor = UIColor(hex: "#5086FF")
-    static let backgroundColor = UIColor.white
-    
-    static let accentColor = UIColor(hex: "#17191C")
-    static let lightGray = UIColor(hex: "#ECEFF2")
-    static let darkGray = UIColor(hex: "#AAAEB6")
-    
-    
-    
-    
+
     // MARK: UIColor with Hex
     public convenience init?(hex: String) {
