+++ Weekand/Weekand/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
