+++ Weekand/Weekand/Resources/Info.plist
