--- Weekand/Weekand/ViewController.swift
@@ -1,18 +1,0 @@
-//
-//  ViewController.swift
-//  Weekand
-//
-//  Created by 이호영 on 2022/05/05.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
